+++ package.json
@@ -5,7 +5,7 @@
   "dependencies": {
     "@babel/core": "^7.12.3",
     "@babel/runtime": "^7.12.1",
-    "@chainsafe/chainbridge-contracts": "^1.0.3",
+    "@chainsafe/chainbridge-contracts": "1.0.5",
     "@chainsafe/web3-context": "1.2.0",
     "@chainsafe/common-components": "1.0.26",
     "@chainsafe/common-theme": "1.0.6",
